# Add `is_*` and `as_*` methods to the event enums