--- a/src/event.rs
+++ b/src/event.rs
@@ -677,7 +677,7 @@ impl Event {
         }
     }
 
-    /// Returns an Option containing the KeyEvent if the event is a key repeat event.
+    /// Returns an Option containing the `KeyEvent` if the event is a key repeat event.
     #[inline]
     pub fn as_key_repeat_event(&self) -> Option<KeyEvent> {
         match self {