@joshka 2024-11-26
There is an [in-flight PR in derive_more](https://github.com/JelteF/derive_more/pull/415) that would be also be useful. It adds an `AsVariant` derive. This would make it easy to deal with e.g. `if let Some(mouse) = event.as_mouse() { ... }` etc. I'd like to consider waiting for that and adding it to this PR.

Obviously a proper match statement is much better when dealing with more than one event type, but there are many use cases that this change would make simpler.