@joshka 2024-12-26
The derive_more change seems stalled for now, so I added the `as_*` functions manually.