@archseer 2025-02-14
I'm against this change. It introduces heavy additional proc macro dependencies that pad out compile time to save a couple lines of code: dependencies like `derive_more` have no place in libraries. It'll increase compile time for us downstream on helix for functions we don't even intend on using.

These types of helpers were already possible with a few lines of user code:

```rust
fn as_key_press(event: &Event) {
  ...
}

if let Some(key) = as_key_press(&event) { ... }
```

At the very least it'd be nice if these methods were behind an optional feature flag (default to on) so we can turn them off.