@the-mikedavis 2025-02-14
Some numbers running locally with a fresh project...

1. `cargo new --bin example`, `cd example`
2. `cargo add --git https://github.com/crossterm-rs/crossterm crossterm --rev <rev>`
    * before this change: `1bcfa9729c8a08b78e236271efe989e766da2926`
    * after: `e063091312d25c21095e69ef547abc00a90ebf7d`
3. `time cargo c`

before:

```
real	0m1.851s
user	0m4.356s
sys	0m1.640s
```

after:

```
real	0m3.539s
user	0m4.650s
sys	0m0.996s
```