@joshka 2025-02-14
Optimizing to avoid crates that are part of almost every 

Quantifying the compile times on an Macbook M2:

Before this change
```
git switch --detach 1bcfa972 
HEAD is now at 1bcfa97 Fix build failure with event-stream, rustix, and use-dev-tty (#955)

cargo clean; time cargo build
...
cargo build  1.95s user 0.59s system 125% cpu 2.016 total
```
After
```
git switch master
Previous HEAD position was 1bcfa97 Fix build failure with event-stream, rustix, and use-dev-tty (#955)
Switched to branch 'master'
Your branch is up to date with 'upstream/master'.

cargo clean ; time cargo build
cargo build  2.51s user 0.75s system 121% cpu 2.688 total
```

572ms extra on a 2 year old laptop doesn't give me any concern about compile times. Help me understand the rationale for your concern here. The procmacro2/syn/quote libraries are already in your dependency tree for helix and the derive_more macro dependency only brings in the macros for the specific derives.

```
cd helix
cargo clean; time cargo build
cargo build  74.48s user 9.25s system 344% cpu 24.286 total
```
(that's on the second build after everything is downloaded).