@the-mikedavis 2025-02-14
Should we eventually remove serde and others that make proc macros necessary then we'd be stuck on tearing this out of crossterm as well.

It's just needless for a library to require these extra dependencies for a cosmetic improvement for UX, and IMO isn't justified (https://github.com/crossterm-rs/crossterm?tab=readme-ov-file#dependency-justification). Libraries should have a small dependency footprint. Adding functions like these sounds good to me but adding a dependency to make it happen isn't necessary.