@joshka 2025-02-14
To quantify doing this as functions, there are 32 variants where these functions would be added, each would be 4 lines of code including doc comment. That's 128 lines of code that this makes not necessary to read / bug check / maintain / ensure is consistent when adding new variants / ...

The tradeoff here is the cost to maintain boilerplate code. I don't generally think it's worth it to save < seconds of compile time to avoid using a library in comparison to the reduction of cognitive load and maintenance burden like the derive macros used here do. That's my specific tradeoff calculus, which I'd apply in general terms in most situations unless there's some compelling argument otherwise.

I'm not the main maintainer here, so I'll defer to Timon on making the final call, but I believe that there's reasonable justification here to use this approach.