@the-mikedavis 2025-02-14
We just have a different tradeoff calculus as you say - we can probably agree to disagree on whether this is justified. I hate to be preachy/normative but for libraries I think it's 'proper"/important to be judicious about dependencies because you're forcing your cost-benefit tradeoff on any consuming libraries and applications. An alternative to reverting or explicitly implementing these dependencies we'd be happy with is gating this behind a feature flag (ideally default=off). As you say though it's up to Timon.