@joshka 2025-02-14
One other part of dependency calculus (though more informative than normative) I generally pay attention to is the relative usages of each library. Derive_more has  6,469,406 downloads per month while crossterm has 1,857,699 downloads per month (and the crate usages. These numbers suggest there's a reasonable amount of safety that comes from choosing to use this library. It's well supported by users across the rust ecosystem and shouldn't generally be avoided.

The rationale for making these on by default is that these patterns are useful for pretty much all applications. As a maintainer of Ratatui, I've seen a heck of a lot of TUI applications that our users have built with the combination this library and ours. These changes would simplify the logic in many many cases. I understand that helix is a popular editor, but it's a single data point in comparison to thes.

Also, I helix is in a situation where the majority of your users install the product from a package manager where the compilation step has already been done. Would that be a correct assumption, or am I missing something? Where specifically are you trying to save compile time?