@the-mikedavis 2025-02-14
I can't speak for @archseer but I'm pretty sure what he's saying about compile time part is more of a general concern about taking dependencies. We already have the proc macro stuff in our tree so that isn't a concrete problem. We take issue with being forced to take unnecessary dependencies transitively. These functions might be useful for the out-of-the-box story for ratatui but the functions are unnecessary for us, and IMO it's unreasonable to pass on the cost of another dependency, no matter how common in the ecosystem, with no way to opt-out, when it's not actually important for core functionality.